Rename "app/core" to "app/domain"
