+++ src/fetchers/ip-fetcher.ts
@@ -16,7 +16,7 @@
  */
 
 import { ConnectionStatus } from 'mysterium-tequilapi'
-import Connection from '../app/core/connection'
+import Connection from '../app/domain/connection'
 import Ip from '../app/models/ip'
 import { ConnectionStatusEnum } from '../libraries/tequil-api/enums'
 import { FetcherBase } from './fetcher-base'
